rename elfedits
In the previous commit, I had not properly removed prebuilts/gcc/ from
manifests upstream.  This failed presubmits in aosp/1441815.

The prefix for the elfedit binaries is indeed CROSS_COMPILE, but for
Android common kernels, that's the tool prefix, not the target triple
(CLANG_TRIPLE).  It seems I myself continue to mix these up.

Bug: 141693040
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Change-Id: I0338c4fcd740866ec93effbe4a4788dd019366cc